requirements.txt
redis==4.6.0
kafka-python==2.0.2 